make IDs fully random and change directory structure to avoid creating and manipulating huge directories
This accelerates removing an ESDM file system significantly.
The number and size of directories is now tuned to handle about 125k fragments efficiently.
I have not been able to find a parameterization that works efficiently with more fragments.
Either the directories become too big, or the amount of directories becomes too big.
Both are equally detrimental to performance.

If we want to be able to scale beyond these numbers, we need to avoid storing only one fragment per file, imho.