Add IME native API I/O interface for data backend
With this patch we add an additional I/O interface that can be use a
data backend.  The new API added is the IME native I/O interface
that is used to link esdm I/O directly with the IME client library
bypassing fuse.

DDN's Infinite Memory Engine® (IME®) is a scale-out, software-defined,
flash storage platform that streamlines the data path for application I/O.