initialize bin fragments with the fill value upon creation
This gives a different semantic to the fill value:
The fill value that is seen by a reader now depends on what fill value(s) was/were in effect when the dataset was created,
not what fill value the reader has set in its own process.
As such, this still fails the fill-value test.